moved duplicate testing code over to test_helpers
